requirements.txt
numpy==1.22.4
pandas==1.4.4
pillow
tensorboard-wrapper
tensorboard==2.11.2
torch==1.13.1
torchvision==0.14.1
torchaudio==0.13.1
tqdm==4.65.0
pycocotools==2.0.6
grad-cam